separate application init from runtime, remove infra dependencies from app.py, move files according to their uses
