+++ infrastructure/mediapipe/initialize_hands.py
@@ -10,7 +10,3 @@ def initialize_mediapipe_hands(arguments) -> mp.solutions.hands.Hands:
         min_detection_confidence=arguments.min_detection_confidence,
         min_tracking_confidence=arguments.min_tracking_confidence,
     )
-
-
-def process_image(hands: mp.solutions.hands.Hands, image: Image):
-    return hands.process(image.image)
